Remove duplication of `VpnAccountSummary`.
