+++ nym-vpn-core/crates/nym-vpn-api-client/src/types/mod.rs
@@ -11,8 +11,7 @@ mod resolver_overrides;
 mod test_fixtures;
 
 pub use account::{
-    Error as AccountError, VpnAccount, VpnAccountMode, VpnAccountSummary, VpnApiTime,
-    VpnApiTimeSynced,
+    Error as AccountError, VpnAccount, VpnAccountMode, VpnApiTime, VpnApiTimeSynced,
 };
 pub use device::{Device, DeviceStatus};
 pub use gateway::{GatewayMinPerformance, GatewayType, ScoreThresholds};
